refactor: reorganize logging and template imports across the application
- Replaced references to shared logging and templates with infrastructure counterparts for consistency.
- Introduced a new logging module to centralize logging configuration and utilities.
- Added a new crypto module for password and token hashing functionalities.
- Updated various services, routes, and middleware to utilize the new logging and template structures, enhancing maintainability and clarity.
