+++ infrastructure/email/zeptomail.py
@@ -12,7 +12,7 @@ from jinja2 import Environment, FileSystemLoader, select_autoescape
 
 from config import EmailSettings
 from infrastructure.http_client import HttpClient
-from shared.logging import get_logger
+from infrastructure.logging import get_logger
 
 log = get_logger(__name__)
