+++ infrastructure/geoip.py
@@ -14,7 +14,7 @@ import geoip2.database
 import geoip2.errors
 import maxminddb
 
-from shared.logging import get_logger
+from infrastructure.logging import get_logger
 
 log = get_logger(__name__)
