+++ main.py
@@ -12,7 +12,7 @@ from dotenv import load_dotenv
 load_dotenv()  # Must run before any module reads os.environ
 
 from app import create_app  # noqa: E402
-from shared.logging import setup_logging  # noqa: E402
+from infrastructure.logging import setup_logging  # noqa: E402
 
 setup_logging()
