+++ repositories/api_key_repository.py
@@ -10,9 +10,9 @@ from __future__ import annotations
 from bson import ObjectId
 from pymongo.errors import PyMongoError
 
+from infrastructure.logging import get_logger
 from repositories.base import BaseRepository
 from schemas.models.api_key import ApiKeyDoc
-from shared.logging import get_logger
 
 log = get_logger(__name__)
