+++ repositories/app_grant_repository.py
@@ -12,9 +12,9 @@ from datetime import datetime, timezone
 from bson import ObjectId
 from pymongo.errors import PyMongoError
 
+from infrastructure.logging import get_logger
 from repositories.base import BaseRepository
 from schemas.models.app_grant import AppGrantDoc
-from shared.logging import get_logger
 
 log = get_logger(__name__)
