+++ repositories/blocked_url_repository.py
@@ -10,8 +10,8 @@ from __future__ import annotations
 
 from pymongo.errors import PyMongoError
 
+from infrastructure.logging import get_logger
 from repositories.base import BaseRepository
-from shared.logging import get_logger
 
 log = get_logger(__name__)
