+++ repositories/click_repository.py
@@ -15,8 +15,8 @@ from typing import Any
 
 from pymongo.errors import PyMongoError
 
+from infrastructure.logging import get_logger
 from repositories.base import BaseRepository
-from shared.logging import get_logger
 
 log = get_logger(__name__)
