+++ repositories/token_repository.py
@@ -14,9 +14,9 @@ from datetime import datetime, timedelta, timezone
 from bson import ObjectId
 from pymongo.errors import PyMongoError
 
+from infrastructure.logging import get_logger
 from repositories.base import BaseRepository
 from schemas.models.token import VerificationTokenDoc
-from shared.logging import get_logger
 
 log = get_logger(__name__)
