+++ services/oauth_service.py
@@ -24,6 +24,7 @@ from bson import ObjectId
 
 from errors import AppError, ConflictError, NotFoundError, ValidationError
 from infrastructure.email.protocol import EmailProvider
+from infrastructure.logging import get_logger
 from repositories.user_repository import UserRepository
 from schemas.dto.responses.auth import OAuthProviderDetail
 from schemas.models.user import (
@@ -38,7 +39,6 @@ from schemas.models.user import (
 )
 from schemas.results import AuthResult
 from services.token_factory import TokenFactory
-from shared.logging import get_logger
 
 log = get_logger(__name__)
