+++ services/stats_service.py
@@ -30,6 +30,7 @@ from errors import (
     NotFoundError,
     ValidationError,
 )
+from infrastructure.logging import get_logger
 from repositories.click_repository import ClickRepository
 from repositories.url_repository import UrlRepository
 from schemas.dto.requests.stats import StatsQuery
@@ -40,7 +41,6 @@ from schemas.enums.stats import (
 )
 from shared.aggregation_strategies import AggregationStrategyFactory
 from shared.datetime_utils import parse_datetime
-from shared.logging import get_logger
 
 log = get_logger(__name__)
