+++ tests/test_utils.py
@@ -2,15 +2,15 @@ import string
 from datetime import datetime, timedelta
 from urllib.parse import unquote
 
-from shared.aggregation_strategies import convert_country_name
-from shared.generators import generate_short_code
-from shared.legacy_helpers import (
+from routes.legacy.helpers import (
     add_missing_dates,
     calculate_click_averages,
     humanize_number,
     is_positive_integer,
     top_four,
 )
+from shared.aggregation_strategies import convert_country_name
+from shared.generators import generate_short_code
 from shared.validators import (
     validate_alias,
     validate_emoji_alias,
